divide pb entity module
