+++ dev/server/grpc/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+relentless-dev-server-grpc-entity = { path = "./entity" }
 prost = "0.13.4"
 tonic = "0.12.3"
 tower = "0.5.2"
@@ -13,5 +14,3 @@ tracing-subscriber = "0.3.19"
 
 num = "0.4.3"
 
-[build-dependencies]
-tonic-build = "0.12.3"
