+++ dev/server/grpc/entity/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "relentless-dev-server-grpc-entity"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tonic = "0.12.3"
+prost = "0.13.4"
+num = "0.4.3"
+
+[build-dependencies]
+tonic-build = "0.12.3"
