+++ dev/server/grpc/examples/client.rs
@@ -1,9 +1,4 @@
-use hello_world::greeter_client::GreeterClient;
-use hello_world::HelloRequest;
-
-pub mod hello_world {
-    tonic::include_proto!("helloworld");
-}
+use relentless_dev_server_grpc_entity::helloworld_pb::{greeter_client::GreeterClient, HelloRequest};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
