+++ dev/server/grpc/src/service/helloworld.rs
@@ -1,11 +1,5 @@
+use relentless_dev_server_grpc_entity::helloworld_pb::{greeter_server::Greeter, HelloReply, HelloRequest};
 use tonic::{Request, Response, Status};
-
-use hello_world::greeter_server::Greeter;
-use hello_world::{HelloReply, HelloRequest};
-
-pub mod hello_world {
-    tonic::include_proto!("helloworld");
-}
 
 #[derive(Debug, Default)]
 pub struct MyGreeter {}
